working on changing ReactQuill editor
