+++ react-app/src/components/Issue/Quill/QuillAdd.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import ReactQuill from 'react-quill';
-import EditorToolbar, { modules, formats } from './EditorToolbar';
+// import ReactQuill from 'react-quill';
+import { QuillEditor } from './QuillEditor';
 import 'react-quill/dist/quill.snow.css';
 import './TextEditor.css';
 
@@ -10,15 +10,16 @@ const QuillAdd = ({ placeholder }) => {
 
     return (
         <div className="quill-add">
-            <EditorToolbar toolbarId={'t1'} />
+            {/* <EditorToolbar toolbarId={'t1'} />
             <ReactQuill
                 theme="snow"
                 value={body}
-                onChange={e => setBody(e.target.value)}
+                onChange={value => setBody(value)}
                 placeholder={placeholder}
                 modules={modules('t1')}
                 formats={formats}
-            />
+            /> */}
+            <QuillEditor />
         </div>
     )
 };
